+++ app/src/main/java/srimani7/apps/feedfly/navigation/NewFeedScreen.kt
@@ -32,7 +32,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import srimani7.apps.feedfly.ui.GroupsPicker
-import srimani7.apps.feedfly.ui.RssItemsColumn
+import srimani7.apps.feedfly.ui.articles.RssItemsColumn
 import srimani7.apps.feedfly.viewmodel.HomeViewModal
 import srimani7.apps.rssparser.ParsingState
