+++ app/src/main/java/srimani7/apps/feedfly/ui/DescriptionText.kt
@@ -4,38 +4,9 @@ package srimani7.apps.feedfly.ui
 
 import android.text.Html.ImageGetter
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.combinedClickable
-import androidx.compose.foundation.gestures.rememberTransformableState
-import androidx.compose.foundation.gestures.transformable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableFloatStateOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.FilterQuality
-import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.window.Dialog
 import androidx.core.text.HtmlCompat
-import coil.compose.AsyncImage
-import coil.request.CachePolicy
-import coil.request.ImageRequest
-import srimani7.apps.feedfly.R
 
 @Composable
 fun HtmlImage(text: String, imageGetter: ImageGetter) {
@@ -44,61 +15,3 @@ fun HtmlImage(text: String, imageGetter: ImageGetter) {
     }
 }
 
-@Composable
-fun ShowImageDialog(
-    imageSrc: String,
-    onDismiss: () -> Unit
-) {
-    var scale by remember { mutableFloatStateOf(1f) }
-    var offset by remember { mutableStateOf(Offset.Zero) }
-    val transformableState =
-        rememberTransformableState { zoomChange, offsetChange, _ ->
-            scale *= zoomChange
-            offset += offsetChange
-        }
-    Dialog(
-        onDismissRequest = onDismiss
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .combinedClickable(
-                    interactionSource = MutableInteractionSource(),
-                    indication = null,
-                    onClick = onDismiss
-                )
-        ) {
-            CircularProgressIndicator(Modifier.align(Alignment.Center))
-            AsyncImage(
-                model = ImageRequest.Builder(LocalContext.current)
-                    .data(imageSrc.replaceFirst("http:", "https:"))
-                    .crossfade(true)
-                    .error(R.drawable.baseline_error_outline_24)
-                    .diskCachePolicy(CachePolicy.DISABLED)
-                    .build(),
-                contentDescription = "image",
-                contentScale = ContentScale.Fit,
-                alignment = Alignment.Center,
-                modifier = Modifier
-                    .fillMaxSize()
-                    .graphicsLayer {
-                        scaleX = scale
-                        scaleY = scale
-                        translationX = offset.x
-                        translationY = offset.y
-                    }
-                    .transformable(transformableState),
-                filterQuality = FilterQuality.High,
-            )
-            IconButton(
-                onClick = onDismiss,
-                modifier = Modifier.align(Alignment.TopEnd)
-            ) {
-                Icon(Icons.Default.Close, "close", tint = Color.White)
-            }
-        }
-    }
-    LaunchedEffect(scale) {
-        if (scale < 0.7f) onDismiss()
-    }
-}
